+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog - math_expressions
 
+### 0.0.2 - 2013-10-xx
+
+- Improve test coverage
+- Improve documentation
+- Fix bug in simplification of Minus
+- Fix bug in simplification of Times
+
 ### 0.0.1+1 - 2013-10-06
 
 - Improve documentation and examples